Problem: not clear why omni_httpd requires omni_sql
omni_httpd doesn't publicly use any of omni_sql interface and therefore
it is confusing why would it need to depend on omni_sql

Solution: extract libomnisql static library
and make omni_httpd use it directly instead of relying on an extension
as it is indeed a bit more of an internal implementation detail, at
least for now.

This also fixes a bug in deparsing where strings were duplicated with
stdup instead of pstrdup (won't work with pfree and can lead to memory
leaks otherwise)
